# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ecosystem Graph API",
    "description": "Query ecosystem — component, contract, dependency, ownership\n\n**ไม่มี endpoint ไหนเขียนข้อมูล** แหล่งความจริงคือ `ecosystem.yaml` การเปลี่ยนแปลงทำผ่าน import เท่านั้น (`/ask` เป็น POST เพราะต้องส่งคำถามใน body ไม่ใช่เพราะเขียนข้อมูล)",
    "version": "0.1.0"
  },
  "paths": {
...
    "/teams/{team_id}": {
      "get": {
        "tags": [
          "team"
        ],
        "summary": "ทีมหนึ่งทีม",
        "operationId": "get_team_teams__team_id__get",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Team Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Team Teams  Team Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/teams/{team_id}/components": {
      "get": {
        "tags": [
          "team"
        ],
        "summary": "ทีมนี้เป็นเจ้าของ component อะไรบ้าง",
        "operationId": "team_components_teams__team_id__components_get",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Team Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Team Components Teams  Team Id  Components Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/components": {
      "get": {
        "tags": [
          "component"
        ],
        "summary": "component ทั้งหมด — กรองได้",
        "operationId": "list_components_components_get",
        "parameters": [
          {
            "name": "team",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "กรองตามทีมเจ้าของ",
              "title": "Team"
            },
            "description": "กรองตามทีมเจ้าของ"
          },
          {
            "name": "plane",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "กรองตาม plane ที่ implement",
              "title": "Plane"
            },
            "description": "กรองตาม plane ที่ implement"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "active / in-development / scaffold / planned",
              "title": "Status"
            },
            "description": "active / in-development / scaffold / planned"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Components Components Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/components/{component_id}": {
      "get": {
        "tags": [
          "component"
        ],
        "summary": "component หนึ่งตัว",
        "operationId": "get_component_components__component_id__get",
        "parameters": [
          {
            "name": "component_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Component Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Component Components  Component Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/components/{component_id}/dependencies": {
      "get": {
        "tags": [
          "graph"
        ],
        "summary": "component นี้ขึ้นกับใครบ้าง (ขาขึ้น)",
        "operationId": "dependencies_components__component_id__dependencies_get",
        "parameters": [
          {
            "name": "component_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Component Id"
            }
          },
          {
            "name": "depth",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10,
              "minimum": 1,
              "default": 10,
              "title": "Depth"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Dependencies Components  Component Id  Dependencies Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/components/{component_id}/dependents": {
      "get": {
        "tags": [
          "graph"
        ],
        "summary": "ใครขึ้นกับ component นี้บ้าง (ขาลง)",
        "operationId": "dependents_components__component_id__dependents_get",
        "parameters": [
          {
            "name": "component_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Component Id"
            }
          },
          {
            "name": "depth",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10,
              "minimum": 1,
              "default": 10,
              "title": "Depth"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Dependents Components  Component Id  Dependents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/contracts/{contract_name}/v{version}": {
      "get": {
        "tags": [
          "contract"
        ],
        "summary": "contract หนึ่งตัว",
        "operationId": "get_contract_contracts__contract_name__v_version__get",
        "parameters": [
          {
            "name": "contract_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Contract Name"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Contract Contracts  Contract Name  V Version  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/contracts/{contract_name}/v{version}/impact": {
      "get": {
        "tags": [
          "graph"
        ],
        "summary": "เปลี่ยน contract นี้แล้วกระทบใคร",
        "operationId": "contract_impact_contracts__contract_name__v_version__impact_get",
        "parameters": [
          {
            "name": "contract_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Contract Name"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Contract Impact Contracts  Contract Name  V Version  Impact Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/planes/{plane_id}": {
      "get": {
        "tags": [
          "plane"
        ],
        "summary": "plane หนึ่งตัว",
        "operationId": "get_plane_planes__plane_id__get",
        "parameters": [
          {
            "name": "plane_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plane Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Plane Planes  Plane Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/repositories/{repo_id}": {
      "get": {
        "tags": [
          "repository"
        ],
        "summary": "repository หนึ่งตัว",
        "operationId": "get_repository_repositories__repo_id__get",
        "parameters": [
          {
            "name": "repo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Repo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Repository Repositories  Repo Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/ask": {
      "post": {
        "tags": [
          "advisor"
        ],
        "summary": "ถามจากมุมของทีม — ไม่เขียนข้อมูลใด ๆ",
        "operationId": "ask_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Ask Ask Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/contracts/{contract_name}/v{version}/coordination": {
      "get": {
        "tags": [
          "advisor"
        ],
        "summary": "เปลี่ยน contract นี้ — ความเสี่ยงและลำดับการประสานงาน",
        "operationId": "coordination_contracts__contract_name__v_version__coordination_get",
        "parameters": [
          {
            "name": "contract_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Contract Name"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version"
            }
          },
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "claude | chatgpt | offline",
              "title": "Provider"
            },
            "description": "claude | chatgpt | offline"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Coordination Contracts  Contract Name  V Version  Coordination Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/work/current": {
      "get": {
        "tags": [
          "github"
        ],
        "summary": "ตอนนี้ใครทำอะไรอยู่ — แยก in-progress ออกจาก declared",
        "operationId": "work_current_work_current_get",
        "parameters": [
          {
            "name": "team",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "กรองเฉพาะทีมนี้",
              "title": "Team"
            },
            "description": "กรองเฉพาะทีมนี้"
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(in-progress|declared)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Work Current Work Current Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AskRequest": {
        "properties": {
          "team": {
            "type": "string",
            "title": "Team",
            "description": "id ของทีมที่ถาม เช่น delivery-team"
          },
          "question": {
            "type": "string",
            "minLength": 3,
            "title": "Question",
            "description": "คำถามภาษาธรรมชาติ"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider",
            "description": "claude | chatgpt | offline — ไม่ใส่ = ตามค่า config"
          },
          "effort": {
            "type": "string",
            "pattern": "^(low|medium|high|xhigh|max)$",
            "title": "Effort",
            "default": "high"
          }
        },
        "type": "object",
        "required": [
          "team",
          "question"
        ],
        "title": "AskRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}